move smt2_format into separate file
