+++ src/solvers/smt2/smt2_parser.cpp
@@ -8,23 +8,9 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include "smt2_parser.h"
 
+#include "smt2_format.h"
+
 #include <util/arith_tools.h>
-
-std::ostream &operator<<(std::ostream &out, const smt2_parsert::smt2_format &f)
-{
-  if(f.type.id() == ID_unsignedbv)
-    out << "(_ BitVec " << to_unsignedbv_type(f.type).get_width() << ')';
-  else if(f.type.id() == ID_bool)
-    out << "Bool";
-  else if(f.type.id() == ID_integer)
-    out << "Int";
-  else if(f.type.id() == ID_real)
-    out << "Real";
-  else
-    out << "? " << f.type.id();
-
-  return out;
-}
 
 void smt2_parsert::command_sequence()
 {
